Squashed commit of the following:

commit 6a89722cbb2d825fb7f61aee8441647c87a14d35
Author: Matt Wilson <152443343+mwsis@users.noreply.github.com>
Date:   Mon May 19 07:49:41 2025 +1000

    Significant multiple improvements (#11)

    * + added **UNIXem.Util.FS** implementation file to src/CMakeLists.txt

    * **UNIXem.Util.FS** : + added `unixem_util_fs_char_is_path_sep()`

    * ~ renamed TODO.txt => TODO.md

    * TODO.md : + more items

    * **test.component.unistd.mkdtemp** / **test.component.unistd.mkdtemp** : ~ minor mods

    * + added **test.unit.util.fs**

    * `glob()` : ~ simplifying refactoring

    * `glob()` : ~ simplifying refactoring

    * ~ preparatory refactoring

    * + added **UNIXem.Util.Str** API

    * + added `unixem_util_fs_get_home_directory()`

    * `glob()` : ~ preparatory refactoring

    * ~ fix

    * `glob()` : ~ refactored in terms of `unixem_util_fs_get_home_directory()`

    * + added and applied `unixem_util_fs_directory_exists()`

    * ~ formatting

    * version

    * ~ MinGW compatibility

    * ~ CMake dependencies fixes

    ---------

+++ test/component/test.component.glob/entry.cpp
@@ -457,7 +457,7 @@ static void TEST_readdir_CWD_FOR_DIRECTORIES_AND_FILES_AND_DOTS()
         std::vector<std::string>        gl_entries;
 
 
-        for (size_t i = 0; gl.gl_pathc != i; ++i)
+        for (int i = 0; gl.gl_pathc != i; ++i)
         {
             // fprintf(stderr, "%d: %s\n", __LINE__, gl.gl_pathv[i]);
 
@@ -513,7 +513,7 @@ static void TEST_readdir_CWD_FOR_DIRECTORIES_AND_DOTS()
         std::vector<std::string>        gl_entries;
 
 
-        for (size_t i = 0; gl.gl_pathc != i; ++i)
+        for (int i = 0; gl.gl_pathc != i; ++i)
         {
             gl_entries.push_back(gl.gl_pathv[i]);
         }
@@ -568,7 +568,7 @@ static void TEST_readdir_HOMEDIR_FOR_DIRECTORIES_AND_FILES_AND_DOTS()
 
         std::vector<std::string>        gl_entries;
 
-        for (size_t i = 0; gl.gl_pathc != i; ++i)
+        for (int i = 0; gl.gl_pathc != i; ++i)
         {
             char const* path = gl.gl_pathv[i];
